build: Add cloud build yaml and pull script
Add a cloudbuild.yml to use google triggers. It will automate deployment
on push to the main branch.

Add a pull.sh script to be executed on remote server. It will pull
from the repository and rebuild static files. Since nginx is already
running it only rebuilds the website. Thus changes nginx will be done
manually.